requirements.txt
dnspython
networkx
pyOpenSSL